+++ apisix/admin/global_rules.lua
@@ -18,7 +18,6 @@ local core = require("apisix.core")
 local schema_plugin = require("apisix.admin.plugins").check_schema
 local type = type
 local tostring = tostring
-local table_util = require("apisix.utils.table-util")
 
 
 local _M = {
@@ -138,7 +137,7 @@ function _M.patch(id, conf, sub_path)
     local node_value = res_old.body.node.value
 
     if sub_path and sub_path ~= "" then
-        local code, err, node_val = table_util.patch(node_value, sub_path, conf);
+        local code, err, node_val = core.table.patch(node_value, sub_path, conf);
         node_value = node_val
         if code then
             return code, err
